Move common schema and tx logic to Table class
